Algo Improvement
Improved the Computer Move, Earlier it was deleting the elements from the left, It was okay but computer move was not ideal. Therefore, Now what computer is doing is finding the smallest element and then removing it. which is kind of optimal move, Human would have also thought of doing that.